requirements.txt
requests
pandas
matplotlib
python-dotenv
psutil
sqlite3